refactor: isolate focus trap DOM helpers
